address more comments + renaming
